Added graphic for Opossum man
fixes #1
A 11th 16px column has been added to the png file and new graphic for Opossum man has been drawn by combining the body of the creature in the 10th column, changing the color and grafting on the head and tail from the existing opossum graphic.
The enlarged png now fits what the text file is expecting and the creatures are displayed correctly.